requirements.txt
# Python backend
python-dotenv
langchain-community
fastapi
uvicorn
pydantic
langgraph
langchain-openai
langchain-core

# Phase 2 — web fetching
httpx
ddgs
praw
feedparser
# Optional JS-heavy extraction fallback:
# crawl4ai
# playwright

# Token counting
tiktoken

# SSE streaming
sse-starlette
